task3_experiment/requirements.txt
pyyaml
faiss-cpu
sentence-transformers
numpy
torch
